# smartestate-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI-powered Real Estate Platform API for Kazakhstan",
        "title": "SmartEstate API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.smartestate.kz/support",
            "email": "support@smartestate.kz"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/targeting/campaigns": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить список всех рекламных кампаний пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Targeting"
                ],
                "summary": "Список кампаний",
                "responses": {
                    "200": {
                        "description": "Список кампаний",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/smartestate_internal_models.Campaign"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка получения кампаний",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создать новую таргетированную рекламную кампанию",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Targeting"
                ],
                "summary": "Создать рекламную кампанию",
                "parameters": [
                    {
                        "description": "Данные кампании",
                        "name": "campaign",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/smartestate_internal_models.Campaign"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Созданная кампания",
                        "schema": {
                            "$ref": "#/definitions/smartestate_internal_models.Campaign"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка создания кампании",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/targeting/campaigns/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить детальную информацию о рекламной кампании",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Targeting"
                ],
                "summary": "Получить кампанию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID кампании",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о кампании",
                        "schema": {
                            "$ref": "#/definitions/smartestate_internal_models.Campaign"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Кампания не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/targeting/campaigns/{id}/launch": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Запустить рекламную кампанию на выбранных площадках",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Targeting"
                ],
                "summary": "Запустить кампанию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID кампании",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Кампания успешно запущена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Кампания не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка запуска кампании",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/targeting/generate-creatives": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Генерировать рекламные креативы с помощью AI для разных платформ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Targeting"
                ],
                "summary": "Генерировать креативы",
                "parameters": [
                    {
                        "description": "Параметры генерации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.GenerateCreativesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сгенерированные креативы для каждой платформы",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка генерации креативов",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_api_handlers.GenerateCreativesRequest": {
            "type": "object",
            "required": [
                "platforms",
                "property_id"
            ],
            "properties": {
                "platforms": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "facebook",
                        "instagram",
                        "google"
                    ]
                },
                "property_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "smartestate_internal_models.Address": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "postal_code": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                }
            }
        },
        "smartestate_internal_models.Campaign": {
            "type": "object",
            "properties": {
                "budget": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "creatives": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/smartestate_internal_models.Creative"
                    }
                },
                "duration_days": {
                    "type": "integer"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metrics": {
                    "$ref": "#/definitions/smartestate_internal_models.CampaignMetrics"
                },
                "name": {
                    "type": "string"
                },
                "platforms": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "property": {
                    "$ref": "#/definitions/smartestate_internal_models.Property"
                },
                "property_id": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "target_audience": {
                    "$ref": "#/definitions/smartestate_internal_models.TargetAudience"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/smartestate_internal_models.User"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "smartestate_internal_models.CampaignMetrics": {
            "type": "object",
            "properties": {
                "clicks": {
                    "type": "integer"
                },
                "conversions": {
                    "type": "integer"
                },
                "cost_per_click": {
                    "type": "number"
                },
                "cost_per_lead": {
                    "type": "number"
                },
                "ctr": {
                    "type": "number"
                },
                "impressions": {
                    "type": "integer"
                },
                "last_updated": {
                    "type": "string"
                },
                "spend": {
                    "type": "number"
                }
            }
        },
        "smartestate_internal_models.ChatContext": {
            "type": "object",
            "properties": {
                "last_intent": {
                    "type": "string"
                },
                "property_preferences": {
                    "type": "object",
                    "additionalProperties": true
                },
                "search_history": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "smartestate_internal_models.ChatMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/smartestate_internal_models.MessageMetadata"
                },
                "role": {
                    "description": "user, assistant, system",
                    "type": "string"
                },
                "session": {
                    "$ref": "#/definitions/smartestate_internal_models.ChatSession"
                },
                "session_id": {
                    "type": "string"
                }
            }
        },
        "smartestate_internal_models.ChatSession": {
            "type": "object",
            "properties": {
                "context": {
                    "$ref": "#/definitions/smartestate_internal_models.ChatContext"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/smartestate_internal_models.ChatMessage"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/smartestate_internal_models.User"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "smartestate_internal_models.Coordinates": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
        "smartestate_internal_models.Creative": {
            "type": "object",
            "properties": {
                "cta": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "headline": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "media_url": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "platform": {
                    "type": "string"
                },
                "type": {
                    "description": "image, video, carousel",
                    "type": "string"
                }
            }
        },
        "smartestate_internal_models.Features": {
            "type": "object",
            "properties": {
                "amenities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "bathrooms": {
                    "type": "integer"
                },
                "bedrooms": {
                    "type": "integer"
                },
                "furnished": {
                    "type": "boolean"
                },
                "has_balcony": {
                    "type": "boolean"
                },
                "has_elevator": {
                    "type": "boolean"
                },
                "has_parking": {
                    "type": "boolean"
                },
                "year_built": {
                    "type": "integer"
                }
            }
        },
        "smartestate_internal_models.MessageMetadata": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "confidence": {
                    "type": "number"
                },
                "extra": {
                    "type": "object",
                    "additionalProperties": true
                },
                "property_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "smartestate_internal_models.Property": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/smartestate_internal_models.Address"
                },
                "area_sqm": {
                    "type": "number"
                },
                "coordinates": {
                    "$ref": "#/definitions/smartestate_internal_models.Coordinates"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "features": {
                    "$ref": "#/definitions/smartestate_internal_models.Features"
                },
                "floor": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "price": {
                    "type": "number"
                },
                "property_type": {
                    "type": "string"
                },
                "rooms": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "total_floors": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/smartestate_internal_models.User"
                },
                "user_id": {
                    "type": "string"
                },
                "views": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/smartestate_internal_models.PropertyView"
                    }
                },
                "vr_tour_url": {
                    "type": "string"
                }
            }
        },
        "smartestate_internal_models.PropertyView": {
            "type": "object",
            "properties": {
                "duration": {
                    "description": "Duration in seconds",
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "ip_address": {
                    "type": "string"
                },
                "property": {
                    "$ref": "#/definitions/smartestate_internal_models.Property"
                },
                "property_id": {
                    "type": "string"
                },
                "referrer": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/smartestate_internal_models.User"
                },
                "user_agent": {
                    "type": "string"
                },
                "user_id": {
                    "description": "Can be null for anonymous views",
                    "type": "string"
                },
                "viewed_at": {
                    "type": "string"
                }
            }
        },
        "smartestate_internal_models.TargetAudience": {
            "type": "object",
            "properties": {
                "age_range": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "behaviors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "custom_params": {
                    "type": "object",
                    "additionalProperties": true
                },
                "gender": {
                    "type": "string"
                },
                "interests": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "locations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "smartestate_internal_models.User": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "campaigns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/smartestate_internal_models.Campaign"
                    }
                },
                "chat_sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/smartestate_internal_models.ChatSession"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "description": "\u003c- изменено",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "properties": {
                    "description": "Relationships - remove constraint tags for initial migration",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/smartestate_internal_models.Property"
                    }
                },
                "role": {
                    "type": "string"
                },
                "subscription_tier": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}